refactor: extract session titles from AgentWorkspace
